revert openssl changes for dead path

+++ src/deps/uws.zig
@@ -2541,8 +2541,8 @@ pub const us_bun_socket_context_options_t = extern struct {
     request_cert: i32 = 0,
     client_renegotiation_limit: u32 = 3,
     client_renegotiation_window: u32 = 600,
-    min_tls_version: u16 = 0,
-    max_tls_version: u16 = 0,
+    min_tls_version: u32 = 0,
+    max_tls_version: u32 = 0,
 };
 
 pub const create_bun_socket_error_t = enum(c_int) {
